♻️ refact: add web forlder
