+++ SnakeTeste/web/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./styles/style.css">
     <title>Snake Game</title>
 </head>
 <body>
@@ -29,9 +29,18 @@
 
 
 <canvas id="game"></canvas>
-    <canvas id="snakeCanvas" width="400" height="400"></canvas>
-    <script src="game.js"></script>
+<canvas id="snakeCanvas" width="400" height="400"></canvas>
 
+<script type="module">
+    document.addEventListener("DOMContentLoaded", function() {
+        import('./game.js').then((module) => {
+            const { startGame, toggleLeaderboard } = module;
+            // Expose functions globally for onclick
+            window.startGame = startGame;
+            window.toggleLeaderboard = toggleLeaderboard
+        });
+    });
+</script>
 
 </body>
 </html>
