+++ SnakeTeste/web/styles/style.css
@@ -24,4 +24,20 @@ body {
     margin: 0 auto; /* Isso centraliza o canvas na página, se necessário */
 }
 
-/* Adicione outros estilos conforme necessário */
+#loader {
+  display: none;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(255, 255, 255, 0.8);
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+#loader img {
+  width: 50px; /* Adjust the size of the loader image */
+  height: 50px; /* Adjust the size of the loader image */
+}
